Added documentation and restructured files.
