+++ src/main.rs
@@ -10,7 +10,7 @@ use ruscii::drawing::Pencil;
 use ruscii::spatial::{Vec2};
 use ruscii::terminal::Color::{Cyan, Red, White};
 use ruscii::terminal::Window;
-use crate::algorithm::name_molecule;
+use crate::groups::name_molecule;
 use crate::input::{input_insert_mode, input_view_mode};
 use crate::spatial::{GridState, Invert};
 use crate::molecule::BondOrder::{Double, Single, Triple};
@@ -20,7 +20,7 @@ mod spatial;
 mod molecule;
 mod macros;
 mod input;
-mod algorithm;
+mod groups;
 mod pointer;
 mod validation;
 mod chain;
